multibody: Move parsing helpers out of the attic
