+++ multibody/multibody_tree/multibody_plant/BUILD.bazel
@@ -114,7 +114,6 @@ drake_cc_googletest(
     name = "multibody_plant_test",
     data = [
         "test/split_pendulum.sdf",
-        "//attic/multibody/shapes:test_models",
         "//manipulation/models/iiwa_description:models",
         "//multibody/benchmarks/acrobot:models",
         "//multibody/multibody_tree/parsing:test_models",
