+++ multibody/multibody_tree/parsing/scene_graph_parser_detail.h
@@ -7,7 +7,7 @@
 
 #include "drake/geometry/scene_graph.h"
 #include "drake/multibody/multibody_tree/multibody_plant/coulomb_friction.h"
-#include "drake/multibody/parsers/package_map.h"
+#include "drake/multibody/multibody_tree/parsing/package_map.h"
 
 namespace drake {
 namespace multibody {
@@ -99,7 +99,7 @@ multibody_plant::CoulombFriction<double> MakeCoulombFrictionFromSdfCollisionOde(
 /// @throws std::runtime_error if the <mesh> tag is present but
 /// missing <uri> or if the file referenced in <uri> can not be found.
 sdf::Visual ResolveVisualUri(const sdf::Visual& original,
-                             const parsers::PackageMap& package_map,
+                             const parsing::PackageMap& package_map,
                              const std::string& root_dir);
 
 }  // namespace detail
