+++ multibody/multibody_tree/parsing/test/package_map_test_packages/package_map_test_package_b/package.xml
