+++ multibody/multibody_tree/parsing/test/package_map_test_packages/package_map_test_package_set/package_map_test_package_c/package.xml
