+++ multibody/multibody_tree/parsing/test/tri_cube.obj
@@ -1,0 +1,26 @@
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 1.000000 -1.000000
+vn 0.0000 -1.0000 0.0000
+vn 0.0000 1.0000 0.0000
+vn 1.0000 -0.0000 0.0000
+vn 0.0000 -0.0000 1.0000
+vn -1.0000 -0.0000 -0.0000
+vn 0.0000 0.0000 -1.0000
+f 1//1 3//1 4//1
+f 8//2 6//2 5//2
+f 5//3 2//3 1//3
+f 6//4 3//4 2//4
+f 3//5 8//5 4//5
+f 1//6 8//6 5//6
+f 1//1 2//1 3//1
+f 8//2 7//2 6//2
+f 5//3 6//3 2//3
+f 6//4 7//4 3//4
+f 3//5 7//5 8//5
+f 1//6 4//6 8//6
